fix(auth): oAuthStore is used before a valid OAuth config is confirmed
