+++ packages/auth/src/providers/cognito/utils/oauth/attemptCompleteOAuthFlow.ts
@@ -12,6 +12,8 @@ import { completeOAuthFlow } from './completeOAuthFlow';
 import { getAuthUserAgentValue } from '../../../../utils';
 import { getRedirectUrl } from './getRedirectUrl';
 import { handleFailure } from './handleFailure';
+import { cognitoUserPoolsTokenProvider } from '../../tokenProvider';
+import { addInflightPromise } from './inflightPromise';
 
 export const attemptCompleteOAuthFlow = async (
 	authConfig: AuthConfig['Cognito']
@@ -31,6 +33,16 @@ export const attemptCompleteOAuthFlow = async (
 	if (!(await oAuthStore.loadOAuthInFlight())) {
 		return;
 	}
+
+	// when there is valid oauth config and there is an inflight oauth flow, try
+	// to block async calls that require fetching tokens before the oauth flow completes
+	// e.g. getCurrentUser, fetchAuthSession etc.
+	cognitoUserPoolsTokenProvider.setWaitForInflightOAuth(
+		() =>
+			new Promise(async (res, _rej) => {
+				addInflightPromise(res);
+			})
+	);
 
 	try {
 		const currentUrl = window.location.href;
